Fix up Kura launcher and rebuild web UI 2 during setup
This change brings a new Kura launcher configuration, based on the
existing Kura Linux launcher, but fixing the path issues so that
the launcher can be started without the need to "select" any resource
in the IDE first.

Also is the Web UI 2 being rebuild during the initial setup phase and
a launcher is provided so that it can easily be rebuilt later on.

In addition the "soda.comm" and "web" bundle are not imported anymore.
The can be imported manually at a later time, but right now, in the
default setup, the cause some trouble because the "soda.comm" library
is not assembled in a way that Eclipse PDE can handle for starting, so
resources are missing and the web UI has the same bundle name so two
conflicting Web UIs exist.

Signed-off-by: Jens Reimann <jreimann@redhat.com>