Fix concurrency issue in command service
The command service was implemented in a way that it did execute all
command written out to a single shell script. This can cause the issue
that two requests issued at the same time would fail.

This change does write out seperate script files using the
Files.createTempFile method. It also does not fork "chmod" and
"dos2unix" since those operations can be performed inside Java which
saves two additional process forks.

Also is the "cd" statement not necessary when the ProcessBuilder
is instructed to change the directory first.

This change also includes a unit test for the command service.

Signed-off-by: Jens Reimann <jreimann@redhat.com>