Fix up the handling of system properties, kura properties and paths
This change creates a new service which does provide an interface
to loading kura properties (based on kura.configuration and
kura.custom.configuration) and provides an interface to get commons
paths like "data" and "configuration".

All locations in Kura where kura.configuration was manually parsed was
replaced by the common service.

The introduction of the service was necessary in order to provide an
abstraction layer to allow future versions to provide their own service
implementations.

For the deployment services a new DeploymentLocation service was added
which provides the location of the deployment packages directory. This
again was necessary on order to decouple the current deployment stuff
from the rest of the core services. Kura should support different ways
of deploying things.

Signed-off-by: Jens Reimann <jreimann@redhat.com>