Refactor Camel router
This change refactors the XML Camel router classes in a way that
new components can be built using a set of different patterns:

 * Create a router component based on the Java builder
 * Create a router component based on custom XML
 * Create a router component as OSGi DS component, configured by XML
   from a property.
   
The `CamelRouter` class is still available for backward compatibility
but is implemented using the refactored classes so that the quickstart
examples still work.

Also does the "kura-cloud" component now pick up the CloudService
from OSGi by default.

Signed-off-by: Jens Reimann <jreimann@redhat.com>