#640: Fix issue initializing modem
This is workaround in order to fix the initialization issue around
SystemService.

Currently the SystemService depends on the NetworkService for
providing network information (which seems not optimal). The network
manager calls back to the core bundle, which is hosting the "linux
utils" (which is not optimal either).

This change does a lazy initialization of the LinuxProcessUtil since
only two methods require the initialization information of this class.

Second the SupportedUsbModems class is implemented now using
commons-exec for calling "lsusb" which may allow to drop the dependency
on the "linux.utils" package or "kura.core" in the future.

This change depends on CQ12140 [1]

[1] https://dev.eclipse.org/ipzilla/show_bug.cgi?id=12140

Signed-off-by: Jens Reimann <jreimann@redhat.com>